very first db schema
